updated routes except telegramHandle

+++ FacilityBooking/facilities.py
@@ -1,9 +1,9 @@
 from flask import Flask, request, make_response
 from flask_cors import CORS, cross_origin
-
 import pymongo
 import json
 from datetime import datetime
+from threading import Thread
 
 
 def removeObjectID(xs):
@@ -25,20 +25,13 @@ myclient = client = pymongo.MongoClient(
     "mongodb+srv://rhdevs-db-admin:rhdevs-admin@cluster0.0urzo.mongodb.net/RHApp?retryWrites=true&w=majority")
 db = myclient["RHApp"]
 
-CROSS_ORIGINS_LIST = "https://rhapp.cjunxiang.vercel.app"
-
 # Flask
 app = Flask("rhapp")
-CORS(app, origins=CROSS_ORIGINS_LIST, headers=['Content-Type'],
-     expose_headers=['Access-Control-Allow-Origin'], supports_credentials=True)
-
-# Session
-session = {}
-# session format : {userID: {
-#                           sessionID: VALUE
-#                           startTime : VALUE
-#                           expiry : VALUE
-#                  }}
+CORS(app, resources={r"/*": {"origins": "*"}})
+# resources allowed to be accessed explicitly
+# response.headers.add("Access-Control-Allow-Origin", "*"), add this to all responses
+# if the cors still now working
+app.config['CORS_HEADERS'] = 'Content-Type'
 
 
 @app.route('/')
@@ -47,116 +40,95 @@ def root_route():
     return 'What up losers'
 
 
-@app.route('/facilities/all/')
-@cross_origin()
+@app.route('/facilities', methods=["GET"])
+@cross_origin(supports_credentials=True)
 def all_facilities():
     try:
-        data = removeObjectID(list(db.Facilities.find()))
-
+        data = removeObjectID(
+            list(db.Facilities.find().sort([("facilityName", 1)])))
+        response = {"status": "success", "data": {"facilities": data}}
     except Exception as e:
         print(e)
-        return {"err": str(e)}, 400
-    response = make_response(json.dumps(
-        list(data), default=lambda o: str(o)), 200)
-    response.headers.add("Access-Control-Allow-Origin", "*")
-    response.headers.add('Access-Control-Allow-Headers', "*")
-    response.headers.add('Access-Control-Allow-Methods', "*")
-    return response
+        return {"err": str(e), "status": "failed"}, 400
+    return make_response(response)
 
 
-@app.route('/facility/<facilityID>')
-@cross_origin()
+@ app.route('/facilities/<facilityID>', methods=["GET"])
+@ cross_origin(supports_credentials=True)
 def get_facility_name(facilityID):
     try:
         data = removeObjectID(
-            list(db.Bookings.find({"facilityID": int(facilityID)})))
+            list(db.Facilities.find({"facilityID": int(facilityID)})))
+        if len(data) == 0:
+            raise Exception("Facility not found")
+        response = {"status": "success", "data": data}
     except Exception as e:
-        return {"err": str(e)}, 400
-    return make_response(json.dumps(list(data), default=lambda o: str(o)), 200)
+        return {"err": str(e), "status": "failed"}, 400
+    return make_response(response)
 
 
-@app.route('/bookings/user/<userID>')
-def user_bookings(userID):
-    try:
-        data = removeObjectID(list(db.Bookings.find({"userID": userID})))
-    except Exception as e:
-        return {"err": str(e)}, 400
-    return make_response(json.dumps(list(data), default=lambda o: str(o)), 200)
-
-
-@app.route('/bookings')
-@cross_origin()
-def get_all_bookings():
-    try:
-        data = removeObjectID(list(db.Bookings.find({})))
-    except Exception as e:
-        return {"err": str(e)}, 400
-    return make_response(json.dumps(list(data), default=lambda o: str(o)), 200)
-
-
-@app.route('/booking/<bookingID>')
-@cross_origin()
+@ app.route('/bookings/<bookingID>', methods=["GET"])
+@ cross_origin(supports_credentials=True)
 def get_one_booking(bookingID):
     try:
         data = removeObjectID(
             list(db.Bookings.find({"bookingID": int(bookingID)})))
+        response = {"status": "success", "data": data}
+        if len(data) == 0:
+            raise Exception("Booking not found")
     except Exception as e:
-        return {"err": str(e)}, 400
-    return make_response(json.dumps(list(data), default=lambda o: str(o)), 200)
+        return {"err": str(e), "status": "failed"}, 400
+    return make_response(response)
 
 
-@app.route('/bookings/facility/<facilityID>/')
+@ app.route('/bookings/user/<userID>', methods=["GET"])
+@ cross_origin(supports_credentials=True)
+def user_bookings(userID):
+    try:
+        data = removeObjectID(list(db.Bookings.find(
+            {"userID": userID}).sort([("startTime", 1)])))
+        response = {"status": "success", "data": data}
+    except Exception as e:
+        return {"err": str(e), "status": "failed"}, 400
+    return make_response(response)
+
+
+@ app.route('/bookings/facility/<facilityID>/', methods=["GET"])
+@ cross_origin(supports_credentials=True)
 def check_bookings(facilityID):
     try:
-        data = removeObjectID(list(db.Bookings.find({"facilityID": int(facilityID),
-                                                     "startTime": {"$gte": int(request.args.get('startDate'))}, "endTime": {"$lte": int(request.args.get('endDate'))}})))
+        data = removeObjectID(list(db.Bookings.find({"facilityID": int(facilityID), "startTime": {"$gte": int(
+            request.args.get('startTime'))}, "endTime": {"$lte": int(request.args.get('endTime'))}}).sort([("startTime", 1)])))
+        response = {"status": "success", "data": data}
     except Exception as e:
-        return {"err": str(e)}, 400
+        return {"err": str(e), "status": "failed"}, 400
 
-    return make_response(json.dumps(list(data), default=lambda o: str(o)), 200)
-
-
-@app.route('/users/telegramID/<userID>')
-@cross_origin(supports_credentials=True)
-def user_telegram(userID):
-    try:
-        profile = db.Profiles.find_one({"userID": userID})
-        telegramHandle = profile.get(
-            'telegramHandle') if profile else "No User Found"
-        data = {"telegramHandle": telegramHandle}
-
-        if (telegramHandle == "No User Found"):
-            return {"err": "No User Found"}, 400
-    except Exception as e:
-        print(e)
-        return {"err": str(e)}, 400
-    return data, 200
+    return make_response(response)
 
 
-@app.route('/bookings', methods=['POST'])
+@ app.route('/bookings', methods=['POST'])
+@ cross_origin(supports_credentials=True)
 def add_booking():
     try:
-        # if request.cookies.get("userID") == list(db.Bookings.find({"bookingID" : bookingID}))[0]['userID'] :
         formData = request.get_json()
         formData["startTime"] = int(formData["startTime"])
         formData["endTime"] = int(formData["endTime"])
 
         formData["facilityID"] = int(formData["facilityID"])
-        formData["ccaID"] = int(formData["ccaID"])
+        if not formData.get("ccaID"):
+            formData["ccaID"] = int(0)
+        else:
+            formData["ccaID"] = int(formData["ccaID"])
 
         if (formData["endTime"] < formData["startTime"]):
             raise Exception("End time eariler than start time")
-        # else:
-        #     return {"err": "Unauthorised Access"}, 401
 
-        # Check for exisiting booking
-        # x1 <= y2 && y1 <= x2 because ranges are well-formed
         conflict = removeObjectID(list(db.Bookings.find({"facilityID": formData.get("facilityID"),
                                                          "endTime": {
-                                                             "$gte": formData.get('startTime')},
-                                                         "startTime": {
-                                                             "$lte": formData.get('endTime')}
-                                                         })))
+            "$gte": formData.get('startTime')},
+            "startTime": {
+            "$lte": formData.get('endTime')}
+        })))
 
         if (len(conflict) != 0):
             raise Exception("Conflict Booking")
@@ -168,55 +140,102 @@ def add_booking():
 
         formData["bookingID"] = newBookingID
         db.Bookings.insert_one(formData)
+        response = {"status": "success"}
 
     except Exception as e:
         print(e)
-        return {"err": str(e)}, 400
+        return {"err": str(e), "status": "failed"}, 400
 
-    return {"message": "Booking Confirmed"}, 200
+    return make_response(response)
 
 
-@app.route('/bookings/<bookingID>', methods=['GET'])
+@ app.route('/bookings/<bookingID>', methods=['GET'])
+@ cross_origin(supports_credentials=True)
 def get_booking(bookingID):
     try:
-        data = removeObjectID(
+        data = listToIndexedDict(
             list(db.Bookings.find({"bookingID": int(bookingID)})))
+        response = {"status": "success", "data": data}
     except Exception as e:
         print(e)
-        return {"err": str(e)}, 400
-    return make_response(json.dumps(list(data), default=lambda o: str(o)), 200)
+        return {"err": str(e), "status": "failed"}, 400
+    return make_response(response)
 
 
-@app.route('/bookings/<bookingID>', methods=['PUT'])
+@ app.route('/bookings/<bookingID>', methods=['PUT'])
+@ cross_origin(supports_credentials=True)
 def edit_booking(bookingID):
     try:
-        # if request.cookies.get("userID") == list(db.Bookings.find({"bookingID" : bookingID}))[0]['userID'] :
-        db.Bookings.update_one({"bookingID": int(bookingID)}, {
-                               "$set": request.get_json()})
+        formData = request.get_json()
+        if not formData.get("ccaID"):
+            formData["ccaID"] = int(0)
+        else:
+            formData["ccaID"] = int(formData["ccaID"])
 
-        # else:
-        #     return {"err": "Unauthorised Access"}, 401
+        data = list(db.Bookings.find({"bookingID": int(bookingID)}))
+
+        if (len(data) == 0):
+            raise Exception("Booking not found")
+
+        db.Bookings.update_one({"bookingID": int(bookingID)}, {
+            "$set": request.get_json()})
+
+        response = {"status": "success"}
     except Exception as e:
         print(e)
-        return {"err": str(e)}, 400
+        return {"err": str(e), "status": "failed"}, 400
 
-    return {"message": "Booking edited"}, 200
+    return make_response(response)
 
 
-@app.route('/bookings/<bookingID>', methods=['DELETE'])
+@ app.route('/bookings/<bookingID>', methods=['DELETE'])
+@ cross_origin(supports_credentials=True)
 def delete_booking(bookingID):
     try:
-        # if request.cookies.get("userID") == list(db.Bookings.find({"bookingID" : bookingID}))[0].get('userID') :
+
+        data = list(db.Bookings.find({"bookingID": int(bookingID)}))
+
+        if (len(data) == 0):
+            raise Exception("Booking not found")
+
         db.Bookings.delete_one({"bookingID": int(bookingID)})
-        # else:
-        #     return {"err": "Unauthorised Access"}, 401
+
+        response = {"status": "success"}
     except Exception as e:
         print(e)
-        return {"err": str(e)}, 400
+        return {"err": str(e), "status": "failed"}, 400
 
-    return {"message": "Booking Deleted"}, 200
+    return make_response(response)
 
 
-if __name__ == '__main__':
-    # app.run(threaded=True, debug=True)
-    app.run('0.0.0.0', port=8080)
+@ app.route('/users/telegramID/<userID>', methods=["GET"])
+@ cross_origin(supports_credentials=True)
+def user_telegram(userID):
+    try:
+        profile = db.Profiles.find_one({"userID": userID})
+        telegramHandle = profile.get(
+            'telegramHandle') if profile else "No User Found"
+        data = {"telegramHandle": telegramHandle}
+
+        if (telegramHandle == "No User Found"):
+            return {"err": "No User Found"}, 400
+    except Exception as e:
+        print(e)
+        return {"err": str(e), "status": "failed"}, 400
+    return data, 200
+
+
+def keep_alive():
+    t = Thread(target=run)
+    t.start()
+
+
+def run():
+    app.run(host='0.0.0.0', port=8080)
+
+
+keep_alive()
+
+# if __name__ == '__main__':
+#   keep_alive();
+#   app.run('0.0.0.0', port=8080)
